+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 - MAVLink code generator.
 - Generate dialects (standard + custom).
 - Package renamed throughout project.
+
+# 0.2.0
+- Added Equatable support to all MAVLink messages.
+- Enabled props for equatable
+- Update ugvcustom dialect using new generator